Move TerminalApp to separate file
