+++ src/interactive/app/mod.rs
@@ -5,6 +5,7 @@ mod handlers;
 pub mod input;
 mod navigation;
 pub mod tree_view;
+pub mod terminal_app;
 
 pub use bytevis::*;
 pub use common::*;
