+++ src/interactive/app/tests/utils.rs
@@ -18,7 +18,7 @@ use std::{
 use tui::backend::TestBackend;
 use tui_react::Terminal;
 
-use crate::interactive::{app::tests::FIXTURE_PATH, TerminalApp};
+use crate::interactive::{app::tests::FIXTURE_PATH, terminal_app::TerminalApp};
 
 pub fn into_keys<'a>(
     codes: impl IntoIterator<Item = KeyCode> + 'a,
